cut distinction between packaged/unpackaged volume
Evepraisal used to provide the group ID and the unpackaged volume for
all items; we used the group ID to figure out if something was a ship
and then looked up its packaged volume in the static data.

Now Evepraisal doesn't provide the group ID or unpackaged volume - it
provides the item ID (typeID) and the packaged volume.

So this commit removes the distinction between packaged and unpackaged
volumes, and just presumes that all ships will be transported in
packaged form.